Generify imagepickercore and prepare video alternative
