+++ mediapicker/progress-proguard.txt
