+++ mediapicker/src/main/AndroidManifest.xml
