+++ mediapicker/src/main/res-public/values-v21/styles.xml
