+++ mediapicker/src/main/res-public/values/colors.xml
