+++ mediapicker/src/main/res-public/values/dimens.xml
