+++ mediapicker/src/main/res-public/values/public.xml
