+++ mediapicker/src/main/res-public/values/styles.xml
