+++ mediapicker/src/main/res/layout-v21/kau_activity_image_picker_overlay.xml
