+++ mediapicker/src/main/res/layout/kau_activity_image_picker.xml
