+++ mediapicker/src/main/res/layout/kau_blurred_imageview.xml
