+++ mediapicker/src/main/res/layout/kau_iitem_image.xml
