+++ mediapicker/src/main/res/layout/kau_iitem_image_basic.xml
