+++ mediapicker/src/main/res/transition-v21/kau_image_enter.xml
