+++ mediapicker/src/main/res/transition-v21/kau_image_exit_bottom.xml
