+++ mediapicker/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="kau_no_images_found">No images found</string>
-    <string name="kau_no_images_selected">No images have been selected</string>
+    <string name="kau_no_items_found">No items found</string>
+    <string name="kau_no_items_selected">No items have been selected</string>
     <string name="kau_blurrable_imageview">Blurrable ImageView</string>
     <string name="kau_no_images_loaded">No images loaded</string>
 </resources>
